Moved API-related functions to own module
