+++ scraper/scraper.py
@@ -4,61 +4,11 @@ APIs, in order to retrieve product data
 """
 import sys
 import click
-import grequests
+import utils.apis as apis
 from scrapy.crawler import CrawlerProcess
-from utils.constants import MercadoLibreConfig as MLC, HEADERS
+from utils.constants import MercadoLibreConfig as MLC
 from utils.spiders import OLXSpider, ColombiaGamerSpider as CGamerSpider
 from utils.constants import OLXConfig as OLX, ColombiaGamerConfig as CGamer
-
-
-def _handle_exception(request, exception):
-    """
-    Exception handler callback function
-    """
-    print('Could not perform the request due to a problem:')
-    print(exception)
-
-
-def send_request(endpoints, country_id = "", category_id = "",
-                 verbose = False):
-    """
-    Attempts to send a request with for the specified list of endpoints.
-    If the endpoints have $SITE_ID and $CATEGORY_ID URL parameters, the 
-    country_id and category_id parameters are respectively required
-
-    This method returns a list of response objects
-    """
-    pending_requests = []
-
-    for endpoint in endpoints:
-        parsed_endpoint = endpoint.replace('$SITE_ID', country_id) \
-                              .replace('$CATEGORY_ID', category_id)
-
-        pending_requests.append(grequests.get(parsed_endpoint,
-                                              headers = HEADERS))
-
-    responses = grequests.map(pending_requests,
-                              exception_handler = _handle_exception)
-
-    for index, response in enumerate(responses):
-        r_url = response.request.url
-
-        if response.status_code == 200:
-            print(f'\n{"*" * 70}')
-            print(f'SUCCESS! Obtained response #{index} for {r_url}\n')
-            if verbose:
-                print(response.json())
-            print(f'{"*" * 70}\n')
-        else:
-            print(f'\n{"*" * 70}')
-            print(f'Problem with the request to {r_url}. ')
-            print(f'Response #{index}:')
-            print(response.status_code)
-            if verbose:
-                print(response.json())
-            print(f'{"*" * 70}\n')
-
-    return responses
 
 
 def _find_exact_among_records(records_collection, match, prop):
@@ -92,16 +42,17 @@ def run(site, verbose):
     The indexes for the sites are:
 
     0: MercadoLibre\n
-    1: [To implement]
+    1: OLX\n
+    2: ColombiaGamer
 
     Windows Use: `python .\scraper\scraper.py --site=<index> [--verbose=<0|1>]`
 
-    Linux/Unix: `python3 ./scraper/scraper.py --site=<index> [--verbose=<0|1>]`    
+    Linux/Unix: `python3 ./scraper/scraper.py --site=<index> [--verbose=<0|1>]`
     """
     if site == 0:
         print(f'Trying to get {MLC.COUNTRY_NAME.value}\'s id')
 
-        country_responses = send_request([MLC.SITES_URL.value])
+        country_responses = apis.send_request([MLC.SITES_URL.value])
         countries = country_responses[0].json()
 
         found_country = _find_exact_among_records(countries,
@@ -112,7 +63,7 @@ def run(site, verbose):
         print(f'{MLC.COUNTRY_NAME.value}\'s id is {country_id}')
         print(f'Trying to get all the categories for country id {country_id}')
 
-        category_reponses = send_request([MLC.CATEGORIES_URL.value],
+        category_reponses = apis.send_request([MLC.CATEGORIES_URL.value],
                                          country_id = country_id)
         categories = category_reponses[0].json()
 
@@ -128,8 +79,8 @@ def run(site, verbose):
               f'site has id: {category_id}')
         print(f'Retrieving all products for {MLC.CATEGORY_NAME.value}')
 
-        product_responses = send_request([MLC.PRODUCTS_URL.value], country_id,
-                                         category_id)
+        product_responses = apis.send_request([MLC.PRODUCTS_URL.value],
+                                              country_id, category_id)
         products = product_responses[0].json()
 
         print(f'{products["paging"]["total"]} items found in this category')
@@ -139,9 +90,7 @@ def run(site, verbose):
 
         playstations = _find_contains_among_records(products['results'],
                                                     'Playstation', 'title')
-
         print(playstations)
-
     elif site == 1:        
         process = CrawlerProcess()
         process.crawl(OLXSpider, start_urls = [OLX.PRODUCTS_URL.value])
